refactor: isolate pcpignore exclusions
